% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{compute_mndwi}
\alias{compute_mndwi}
\title{Compute MNDWI}
\usage{
compute_mndwi(x, green = "B03", swir1 = "B11")
}
\arguments{
\item{x}{A \code{terra::SpatRaster}.}

\item{green}{Name of green band.}

\item{swir1}{Name of SWIR1 band.}
}
\value{
A single-layer \code{terra::SpatRaster}.
}
\description{
Modified NDWI: \code{(green - swir1) / (green + swir1)}.
}
