% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{compute_nbr}
\alias{compute_nbr}
\title{Compute NBR}
\usage{
compute_nbr(x, nir = "B08", swir2 = "B12")
}
\arguments{
\item{x}{A \code{terra::SpatRaster}.}

\item{nir}{Name of near-infrared band.}

\item{swir2}{Name of SWIR2 band.}
}
\value{
A single-layer \code{terra::SpatRaster}.
}
\description{
Normalized Burn Ratio: \code{(nir - swir2) / (nir + swir2)}.
}
