% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{compute_ndvi}
\alias{compute_ndvi}
\title{Compute NDVI}
\usage{
compute_ndvi(x, nir = "B08", red = "B04")
}
\arguments{
\item{x}{A \code{terra::SpatRaster}.}

\item{nir}{Name of near-infrared band.}

\item{red}{Name of red band.}
}
\value{
A single-layer \code{terra::SpatRaster}.
}
\description{
Compute NDVI
}
