% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{compute_ndwi}
\alias{compute_ndwi}
\title{Compute NDWI}
\usage{
compute_ndwi(x, green = "B03", nir = "B08")
}
\arguments{
\item{x}{A \code{terra::SpatRaster}.}

\item{green}{Name of green band.}

\item{nir}{Name of near-infrared band.}
}
\value{
A single-layer \code{terra::SpatRaster}.
}
\description{
McFeeters-style NDWI: \code{(green - nir) / (green + nir)}.
}
