% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{compute_normalized_difference}
\alias{compute_normalized_difference}
\title{Compute a normalized difference index}
\usage{
compute_normalized_difference(x, band1, band2, name = "ND")
}
\arguments{
\item{x}{A \code{terra::SpatRaster}.}

\item{band1}{Name of first band.}

\item{band2}{Name of second band.}

\item{name}{Name of the output layer.}
}
\value{
A single-layer \code{terra::SpatRaster}.
}
\description{
Compute a normalized difference index
}
