% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_percentile_thresholds.R
\name{compute_percentile_thresholds}
\alias{compute_percentile_thresholds}
\title{Compute percentile thresholds}
\usage{
compute_percentile_thresholds(values, bands, lower_p = 5, upper_p = 95)
}
\arguments{
\item{values}{A data.frame with sampled raster values.}

\item{bands}{Character vector of band names.}

\item{lower_p}{Named numeric vector or single numeric value. Lower percentile.}

\item{upper_p}{Named numeric vector or single numeric value. Upper percentile.}
}
\value{
A data.frame with columns \code{lower_p}, \code{upper_p}, \code{lower}, and \code{upper}.
}
\description{
Computes lower and upper percentile thresholds for one or more bands.
}
