% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_treshholds.R
\name{mask_band_thresholds}
\alias{mask_band_thresholds}
\title{Apply band-wise threshold mask}
\usage{
mask_band_thresholds(
  x,
  bands,
  thresholds,
  operators = rep(">", length(bands)),
  combine = "and",
  keep_non_masked = FALSE
)
}
\arguments{
\item{x}{A \code{terra::SpatRaster}.}

\item{bands}{Character vector of band names.}

\item{thresholds}{Named numeric vector with one threshold per band.}

\item{operators}{Character vector of threshold operators. One of ">", ">=", "<", "<=".}

\item{combine}{How to combine masks. Either "and" or "or".}

\item{keep_non_masked}{Logical. If \code{FALSE}, pixels matching the condition are \code{TRUE}.}
}
\value{
A logical \code{terra::SpatRaster}.
}
\description{
Creates a logical mask from one or more raster bands and thresholds.
}
