% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_fraction.R
\name{mask_fraction}
\alias{mask_fraction}
\title{Calculate mask fraction}
\usage{
mask_fraction(mask)
}
\arguments{
\item{mask}{A logical \code{terra::SpatRaster}.}
}
\value{
Numeric value between 0 and 1.
}
\description{
Calculates the fraction of TRUE pixels in a logical raster mask.
}
