% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_threshold_hist.R
\name{plot_band_threshold_hist}
\alias{plot_band_threshold_hist}
\title{Plot threshold histograms for raster bands}
\usage{
plot_band_threshold_hist(values, bands, thresholds)
}
\arguments{
\item{values}{A data.frame of sampled raster values.}

\item{bands}{Character vector of band names.}

\item{thresholds}{Named numeric vector of thresholds.}
}
\value{
A ggplot object.
}
\description{
Plot threshold histograms for raster bands
}
