% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_percentile_threshold_hist.R
\name{plot_percentile_threshold_hist}
\alias{plot_percentile_threshold_hist}
\title{Plot histograms with percentile thresholds}
\usage{
plot_percentile_threshold_hist(
  values,
  bands,
  params,
  side = c("upper", "lower", "outside", "inside")
)
}
\arguments{
\item{values}{A data.frame with sampled raster values.}

\item{bands}{Character vector of band names.}

\item{params}{A data.frame with columns \code{lower_p}, \code{upper_p}, \code{lower}, and \code{upper}.}

\item{side}{Which threshold side is currently used.}
}
\value{
A ggplot object.
}
\description{
Plot histograms with percentile thresholds
}
