% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_range_threshold_hist.R
\name{plot_range_threshold_hist}
\alias{plot_range_threshold_hist}
\title{Plot histograms with range thresholds}
\usage{
plot_range_threshold_hist(values, bands, ranges)
}
\arguments{
\item{values}{A data.frame with sampled raster values.}

\item{bands}{Character vector of band names.}

\item{ranges}{A matrix or data.frame with columns \code{min} and \code{max}.}
}
\value{
A ggplot object.
}
\description{
Plot histograms with range thresholds
}
