% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_values.R
\name{sample_band_values}
\alias{sample_band_values}
\title{Sample raster values}
\usage{
sample_band_values(x, bands, sample_size = 20000, seed = 1)
}
\arguments{
\item{x}{A \code{terra::SpatRaster}.}

\item{bands}{Character vector of band names.}

\item{sample_size}{Number of pixels to sample.}

\item{seed}{Optional random seed.}
}
\value{
A data.frame with sampled values.
}
\description{
Randomly samples values from selected raster bands.
}
